统计平均每周漏检数、平均每月漏检数，保留两位小数，四舍五入

+++ aiurt-module-patrol/aiurt-module-patrol-biz/src/main/java/com/aiurt/boot/report/model/PatrolReport.java
@@ -58,7 +58,7 @@ public class PatrolReport {
      */
     @ApiModelProperty("漏巡视数")
     @Excel(name = "漏巡视数", width = 15)
-    private Integer missInspectedNumber;
+    private float missInspectedNumber;
     /**
      * 平均每周漏巡视数
      */
